feat: remove ura flag
