--- src/featureFlags/flags/unifiedRouter.ts
@@ -1,11 +1,0 @@
-import { BaseVariant, FeatureFlag, useBaseFlag } from '../index'
-
-export function useRoutingAPIV2Flag(): BaseVariant {
-  return useBaseFlag(FeatureFlag.uraEnabled)
-}
-
-export function useRoutingAPIV2Enabled(): boolean {
-  return useRoutingAPIV2Flag() === BaseVariant.Enabled
-}
-
-export { BaseVariant as UnifiedRouterVariant }
