+++ src/state/index.ts
@@ -6,8 +6,7 @@ import { isTestEnv } from 'utils/env'
 import { updateVersion } from './global/actions'
 import { sentryEnhancer } from './logging'
 import reducer from './reducer'
-import { routingApi } from './routing/slice'
-import { routingApiV2 } from './routing/v2Slice'
+import { routingApiV2 } from './routing/slice'
 
 const PERSISTED_KEYS: string[] = ['user', 'transactions', 'lists']
 
@@ -21,10 +20,9 @@ const store = configureStore({
         // meta.arg and meta.baseQueryMeta are defaults. payload.trade is a nonserializable return value, but that's ok
         // because we are not adding it into any persisted store that requires serialization (e.g. localStorage)
         ignoredActionPaths: ['meta.arg', 'meta.baseQueryMeta', 'payload.trade'],
-        ignoredPaths: [routingApi.reducerPath, routingApiV2.reducerPath],
+        ignoredPaths: [routingApiV2.reducerPath],
       },
     })
-      .concat(routingApi.middleware)
       .concat(routingApiV2.middleware)
       .concat(save({ states: PERSISTED_KEYS, debounce: 1000 })),
   preloadedState: load({ states: PERSISTED_KEYS, disableWarnings: isTestEnv() }),
