+++ src/state/routing/slice.ts
@@ -4,29 +4,10 @@ import { TradeType } from '@uniswap/sdk-core'
 import { ChainId } from '@uniswap/smart-order-router'
 import { getClientSideQuote } from 'lib/hooks/routing/clientSideSmartOrderRouter'
 import ms from 'ms.macro'
-import qs from 'qs'
 import { trace } from 'tracing/trace'
 
-import { QuoteData, QuoteMethod, TradeResult } from './types'
+import { QuoteMethod, QuoteReponse, QuoteState, TradeResult } from './types'
 import { getRouter, isExactInput, shouldUseAPIRouter, transformRoutesToTrade } from './utils'
-
-export enum RouterPreference {
-  AUTO = 'auto',
-  API = 'api',
-  CLIENT = 'client',
-}
-
-// This is excluded from `RouterPreference` enum because it's only used
-// internally for token -> USDC trades to get a USD value.
-export const INTERNAL_ROUTER_PREFERENCE_PRICE = 'price' as const
-
-// routing API quote params: https://github.com/Uniswap/routing-api/blob/main/lib/handlers/quote/schema/quote-schema.ts
-const API_QUERY_PARAMS = {
-  protocols: 'v2,v3,mixed',
-}
-const CLIENT_PARAMS = {
-  protocols: [Protocol.V2, Protocol.V3, Protocol.MIXED],
-}
 
 export interface GetQuoteArgs {
   tokenInAddress: string
@@ -43,21 +24,36 @@ export interface GetQuoteArgs {
   isRoutingAPIPrice?: boolean
 }
 
-enum QuoteState {
-  SUCCESS = 'Success',
-  NOT_FOUND = 'Not found',
+export enum RouterPreference {
+  AUTO = 'auto',
+  API = 'api',
+  CLIENT = 'client',
 }
 
-export const routingApi = createApi({
-  reducerPath: 'routingApi',
+// This is excluded from `RouterPreference` enum because it's only used
+// internally for token -> USDC trades to get a USD value.
+export const INTERNAL_ROUTER_PREFERENCE_PRICE = 'price' as const
+
+const CLIENT_PARAMS = {
+  protocols: [Protocol.V2, Protocol.V3, Protocol.MIXED],
+}
+
+// routing API quote query params: https://github.com/Uniswap/routing-api/blob/main/lib/handlers/quote/schema/quote-schema.ts
+const CLASSIC_SWAP_QUERY_PARAMS = {
+  ...CLIENT_PARAMS,
+  routingType: 'CLASSIC',
+}
+
+export const routingApiV2 = createApi({
+  reducerPath: 'routingApiV2',
   baseQuery: fetchBaseQuery({
-    baseUrl: 'https://api.uniswap.org/v1/',
+    baseUrl: 'https://api.uniswap.org/v2/',
   }),
   endpoints: (build) => ({
     getQuote: build.query<TradeResult, GetQuoteArgs>({
       async onQueryStarted(args: GetQuoteArgs, { queryFulfilled }) {
         trace(
-          'quote',
+          'quote-v2',
           async ({ setTraceError, setTraceStatus }) => {
             try {
               await queryFulfilled
@@ -83,28 +79,39 @@ export const routingApi = createApi({
           }
         )
       },
-      async queryFn(args, _api, _extraOptions, fetch) {
-        const fellBack = false
+      async queryFn(args: GetQuoteArgs, _api, _extraOptions, fetch) {
+        let fellBack = false
         if (shouldUseAPIRouter(args)) {
+          fellBack = true
           try {
             const { tokenInAddress, tokenInChainId, tokenOutAddress, tokenOutChainId, amount, tradeType } = args
-            const type = isExactInput(tradeType) ? 'exactIn' : 'exactOut'
-            const query = qs.stringify({
-              ...API_QUERY_PARAMS,
-              tokenInAddress,
+            const type = isExactInput(tradeType) ? 'EXACT_INPUT' : 'EXACT_OUTPUT'
+
+            const requestBody = {
               tokenInChainId,
-              tokenOutAddress,
+              tokenIn: tokenInAddress,
               tokenOutChainId,
+              tokenOut: tokenOutAddress,
               amount,
               type,
+              configs: [CLASSIC_SWAP_QUERY_PARAMS],
+            }
+
+            const response = await fetch({
+              method: 'POST',
+              url: '/quote',
+              body: JSON.stringify(requestBody),
             })
-            const response = await fetch(`quote?${query}`)
+
             if (response.error) {
               try {
                 // cast as any here because we do a runtime check on it being an object before indexing into .errorCode
                 const errorData = response.error.data as any
                 // NO_ROUTE should be treated as a valid response to prevent retries.
-                if (typeof errorData === 'object' && errorData?.errorCode === 'NO_ROUTE') {
+                if (
+                  typeof errorData === 'object' &&
+                  (errorData?.errorCode === 'NO_ROUTE' || errorData?.detail === 'No quotes available')
+                ) {
                   return { data: { state: QuoteState.NOT_FOUND } }
                 }
               } catch {
@@ -112,12 +119,13 @@ export const routingApi = createApi({
               }
             }
 
-            const quoteData = response.data as QuoteData
-            const tradeResult = transformRoutesToTrade(args, quoteData)
+            const quoteData = response.data as QuoteReponse
+            const tradeResult = transformRoutesToTrade(args, quoteData.quote)
+
             return { data: { ...tradeResult, method: QuoteMethod.ROUTING_API } }
           } catch (error: any) {
             console.warn(
-              `GetQuote failed on routing API, falling back to client: ${error?.message ?? error?.detail ?? error}`
+              `GetQuote failed on API v2, falling back to client: ${error?.message ?? error?.detail ?? error}`
             )
           }
         }
@@ -136,11 +144,8 @@ export const routingApi = createApi({
         }
       },
       keepUnusedDataFor: ms`10s`,
-      extraOptions: {
-        maxRetries: 0,
-      },
     }),
   }),
 })
 
-export const { useGetQuoteQuery } = routingApi
+export const { useGetQuoteQuery } = routingApiV2
