+++ src/state/routing/useRoutingAPITrade.ts
@@ -3,12 +3,11 @@ import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
 import { IMetric, MetricLoggerUnit, setGlobalMetric } from '@uniswap/smart-order-router'
 import { sendTiming } from 'components/analytics'
 import { AVERAGE_L1_BLOCK_TIME } from 'constants/chainInfo'
-import { useRoutingAPIV2Enabled } from 'featureFlags/flags/unifiedRouter'
 import { useRoutingAPIArguments } from 'lib/hooks/routing/useRoutingAPIArguments'
 import ms from 'ms.macro'
 import { useMemo } from 'react'
-import { INTERNAL_ROUTER_PREFERENCE_PRICE, RouterPreference, useGetQuoteQuery } from 'state/routing/slice'
-import { useGetQuoteQuery as useGetQuoteQueryV2 } from 'state/routing/v2Slice'
+import { INTERNAL_ROUTER_PREFERENCE_PRICE, RouterPreference } from 'state/routing/slice'
+import { useGetQuoteQuery as useGetQuoteQueryV2 } from 'state/routing/slice'
 
 import { InterfaceTrade, QuoteMethod, QuoteState, TradeState } from './types'
 
@@ -48,34 +47,16 @@ export function useRoutingAPITrade<TTradeType extends TradeType>(
     routerPreference,
   })
 
-  const shouldUseRoutingApiV2 = useRoutingAPIV2Enabled()
-
   const {
-    isError: isLegacyAPIError,
-    data: legacyAPITradeResult,
-    currentData: currentLegacyAPITradeResult,
-  } = useGetQuoteQuery(shouldUseRoutingApiV2 ? skipToken : queryArgs ?? skipToken, {
+    isError,
+    data: tradeResult,
+    currentData: currentTradeResult,
+  } = useGetQuoteQueryV2(queryArgs ?? skipToken, {
     // Price-fetching is informational and costly, so it's done less frequently.
     pollingInterval: routerPreference === INTERNAL_ROUTER_PREFERENCE_PRICE ? ms`1m` : AVERAGE_L1_BLOCK_TIME,
     // If latest quote from cache was fetched > 2m ago, instantly repoll for another instead of waiting for next poll period
     refetchOnMountOrArgChange: 2 * 60,
   })
-
-  const {
-    isError: isV2APIError,
-    data: v2TradeResult,
-    currentData: currentV2TradeResult,
-  } = useGetQuoteQueryV2(!shouldUseRoutingApiV2 ? skipToken : queryArgs ?? skipToken, {
-    // Price-fetching is informational and costly, so it's done less frequently.
-    pollingInterval: routerPreference === INTERNAL_ROUTER_PREFERENCE_PRICE ? ms`1m` : AVERAGE_L1_BLOCK_TIME,
-    // If latest quote from cache was fetched > 2m ago, instantly repoll for another instead of waiting for next poll period
-    refetchOnMountOrArgChange: 2 * 60,
-  })
-
-  const [tradeResult, currentTradeResult, isError] = shouldUseRoutingApiV2
-    ? [v2TradeResult, currentV2TradeResult, isV2APIError]
-    : [legacyAPITradeResult, currentLegacyAPITradeResult, isLegacyAPIError]
-
   const isCurrent = currentTradeResult === tradeResult
 
   return useMemo(() => {
